Update : firebase db에 user데이터 추가 - 로그인 한 유저 데이터 & 서비스에서 필요한 데이터 추가 - 유저 타입은 usedType.ts파일로 이동
