+++ src/types/usedType.ts
@@ -1,3 +1,5 @@
+import { User } from "firebase/auth";
+
 // 중고 데이터 타입
 export interface MyUsedItemType {
   id?: string;
@@ -72,19 +74,19 @@ export interface Orders {
 
 export interface Sales {
   salesId: string;
-  salesItems:
-    | {
-        itemId: string;
-        itemName: string;
-        price: number;
-        size: string;
-        imageUrl: string;
-        quantity: number;
-        createdAt: string;
-        isSales: boolean;
-        options: string[];
-      }[]
-    | [];
+  salesItems: SalesItemsType[];
+  // | {
+  //     itemId: string;
+  //     itemName: string;
+  //     price: number;
+  //     size: string;
+  //     imageUrl: string;
+  //     quantity: number;
+  //     createdAt: string;
+  //     isSales: boolean;
+  //     options: string[];
+  //   }[]
+  // | [];
 }
 
 export interface Carts {
@@ -108,3 +110,67 @@ interface Wishlists {
   imageUrl: string;
 }
 
+// firebase 유저 타입
+export interface FirebaseUserType extends User {
+  uid: string;
+  email: string | null;
+  displayName: string | null;
+  photoURL: string | null;
+  emailVerified: boolean;
+  phoneNumber: string | null;
+}
+
+interface UsedItems {
+  itemId: string;
+  itemName: string;
+  price: number;
+  size: string;
+  imageUrl: string;
+  quantity: number;
+}
+
+interface SalesItemsType extends UsedItems {
+  createdAt: string;
+  isSales: boolean;
+  options: string[];
+}
+
+interface OrderItemsType {
+  orderId: string;
+  totalPrice: number;
+  items: UsedItems[];
+  buyerInfo: {
+    name: string;
+    address: string;
+    phone: string;
+  };
+  paymentMethod: string;
+  orderDate: string;
+}
+
+interface WishItemsType {
+  itemId: string;
+  itemName: string;
+  price: number;
+  imageUrl: string;
+}
+
+export interface UserDataType {
+  userId: string;
+  userEmail: string | null;
+  userName: string | null;
+  nickname: string | null;
+  userAvatar: string | null;
+  address: string;
+  phone: string | null;
+  orders: OrderItemsType[];
+  sales: {
+    saleId: string;
+    salesItems: SalesItemsType[];
+  };
+  carts: {
+    cartId: string;
+    cartsItems: UsedItems[];
+  };
+  wishlists: WishItemsType[];
+}
